item: virtual item* constructor, became interface
Now Item class will not be used by itself,
it will only be instanciated if a child class
do so. It has a virtual method that is equal to 0,
so the child classes need to implement it. This
implementation on the children classes is a return
for an allocation in the heap of the own class using
the *this.